+++ src/main/java/com/hyz/safebank/service/impl/SavingAccountServiceImpl.java
@@ -11,7 +11,9 @@ import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Optional;
+import java.util.Random;
 
 @Service
 public class SavingAccountServiceImpl implements SavingAccountService{
@@ -41,7 +43,7 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                 .openDate(LocalDate.now())
                 .accType("SAVING")
                 .accountBalance(BigDecimal.ZERO)
-                .interestRate(BigDecimal.ZERO)
+                .interestRate(BigDecimal.valueOf(new Random().nextDouble() * 0.03 + 0.03).setScale(2, BigDecimal.ROUND_HALF_UP))
                 .build();
 
         SavingAccount savedSavingAccount = savingAccountRepository.save(savingAccount);
@@ -56,6 +58,7 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                         .accType(savedSavingAccount.getAccType())
                         .accountBalance(savedSavingAccount.getAccountBalance())
                         .interestRate(savedSavingAccount.getInterestRate())
+                        .customerId(savedSavingAccount.getCustomer().getId())
                                 .build())
                 .build();
     }
@@ -83,6 +86,7 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                         .accType(savingAccount.get().getAccType())
                         .accountBalance(savingAccount.get().getAccountBalance())
                         .interestRate(savingAccount.get().getInterestRate())
+                        .customerId(customerId)
                         .build())
                 .build();
     }
@@ -97,7 +101,16 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                     .build();
         }
 
+        if (depositWithdrawRequest.getAmount().compareTo(BigDecimal.ZERO) <= 0) {
+            return BankResponse.builder()
+                    .responseCode("001")
+                    .responseMessage("Deposit amount must be greater than zero")
+                    .savingAccountInfo(null)
+                    .build();
+        }
+
         Optional<SavingAccount> savingAccount = savingAccountRepository.findById(depositWithdrawRequest.getAccountId());
+
         SavingAccount theSavingAccount = savingAccount.get();
         BigDecimal currentBalance = theSavingAccount.getAccountBalance();
         BigDecimal depositAmount = depositWithdrawRequest.getAmount();
@@ -116,6 +129,7 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                         .accType(theSavingAccount.getAccType())
                         .accountBalance(theSavingAccount.getAccountBalance())
                         .interestRate(theSavingAccount.getInterestRate())
+                        .customerId(theSavingAccount.getCustomer().getId())
                         .build())
                 .build();
 
@@ -131,7 +145,24 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                     .build();
         }
 
+        if (depositWithdrawRequest.getAmount().compareTo(BigDecimal.ZERO) <= 0) {
+            return BankResponse.builder()
+                    .responseCode("001")
+                    .responseMessage("Withdraw amount must be greater than zero")
+                    .savingAccountInfo(null)
+                    .build();
+        }
+
+
         Optional<SavingAccount> savingAccount = savingAccountRepository.findById(depositWithdrawRequest.getAccountId());
+
+        if (savingAccount.get().getAccountBalance().compareTo(depositWithdrawRequest.getAmount()) < 0) {
+            return BankResponse.builder()
+                    .responseCode("001")
+                    .responseMessage("Insufficient balance")
+                    .savingAccountInfo(null)
+                    .build();
+        }
         SavingAccount theSavingAccount = savingAccount.get();
         BigDecimal currentBalance = theSavingAccount.getAccountBalance();
         BigDecimal withdrawAmount = depositWithdrawRequest.getAmount();
@@ -150,6 +181,7 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                         .accType(theSavingAccount.getAccType())
                         .accountBalance(theSavingAccount.getAccountBalance())
                         .interestRate(theSavingAccount.getInterestRate())
+                        .customerId(theSavingAccount.getCustomer().getId())
                         .build())
                 .build();
     }
@@ -164,11 +196,44 @@ public class SavingAccountServiceImpl implements SavingAccountService{
                     .build();
         }
 
+        //check if account has balance
+        Optional<SavingAccount> savingAccount = savingAccountRepository.findById(enquiryRequest.getAccountId());
+        SavingAccount theSavingAccount = savingAccount.get();
+        if (theSavingAccount.getAccountBalance().compareTo(BigDecimal.ZERO) > 0) {
+            return BankResponse.builder()
+                    .responseCode("001")
+                    .responseMessage("Account has balance. Withdraw balance before deleting account")
+                    .savingAccountInfo(null)
+                    .build();
+        }
+
         savingAccountRepository.deleteById(enquiryRequest.getAccountId());
         return BankResponse.builder()
                 .responseCode("002")
                 .responseMessage("Account deleted")
                 .savingAccountInfo(null)
+                .build();
+    }
+
+    @Override
+    public List<SavingAccountInfo> getAllSavingAccounts() {
+        List<SavingAccount> savingAccounts = savingAccountRepository.findAll();
+        return savingAccounts.stream()
+                .map(this::convertToInfo)
+                .toList();
+    }
+
+    @Override
+    public SavingAccountInfo convertToInfo(SavingAccount savingAccount) {
+        return SavingAccountInfo.builder()
+                .accountId(savingAccount.getId())
+                .accountNumber(savingAccount.getAccountNumber())
+                .accountName(savingAccount.getAccountName())
+                .openDate(savingAccount.getOpenDate())
+                .accType(savingAccount.getAccType())
+                .accountBalance(savingAccount.getAccountBalance())
+                .interestRate(savingAccount.getInterestRate())
+                .customerId(savingAccount.getCustomer().getId())
                 .build();
     }
